+++ DotNetProject8/Models/AppointmentStatusResponse.cs
@@ -2,7 +2,7 @@
 {
     public class AppointmentStatusResponse
     {
-        public int AppointmentId { get; set; }
+        public long AppointmentId { get; set; }
         public DateTime AppointmentDate { get; set; } = new DateTime();
         public DateTime AppointmentTime { get; set; } = new DateTime();
         public string? Status { get; set; }
